auth/requirements.txt
aioredis==2.0.0
aiosmtplib==1.1.6
anyio==3.3.1
asgiref==3.4.1
async-timeout==3.0.1
autopep8==1.5.7
blinker==1.4
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
cryptography==3.4.7
dnspython==2.1.0
email-validator==1.1.3
fakeredis==1.6.1
fastapi==0.68.1
fastapi-mail==0.3.7
h11==0.12.0
httpcore==0.13.6
httpx==0.19.0
idna==3.2
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==21.0
Pillow==8.2.0
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
PyJWT==1.7.1
pymongo==3.11.4
pyparsing==2.4.7
python-dotenv==0.19.0
passlib==1.7.4
python-multipart==0.0.5
redis==3.5.3
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.2
uvicorn==0.15.0
